--- a/pandas/core/arrays/string_.py
+++ b/pandas/core/arrays/string_.py
@@ -318,9 +318,7 @@ class StringArray(BaseStringArray, PandasArray):
 
     def _validate(self):
         """Validate that we only store NA or strings."""
-        if len(self._ndarray) and not lib.is_string_array(
-            self._ndarray.ravel("K"), skipna=True
-        ):
+        if len(self._ndarray) and not lib.is_string_array(self._ndarray, skipna=True):
             raise ValueError("StringArray requires a sequence of strings or pandas.NA")
         if self._ndarray.dtype != "object":
             raise ValueError(